requirements.txt
elasticsearch
little_pger
PyPDF2==1.26.0
pytesseract==0.3.0
opencv-python==4.1.1.26
scipy==1.3.1
wheel==0.33.4
numpy==1.17.3
scikit-learn==0.21.3
requests==2.22.0
nltk==3.4.3
stop-words
configparser==3.7.4
Flask==1.0.3
pandas==0.25.2
cmake==3.15.3
gensim==3.8.1
flask_restful==0.3.7
Pillow==6.2.1
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
Werkzeug==0.16.0
PyMuPDF==1.16.5
smart_open==1.8.4
jinja2==2.10.3
markupsafe==1.1.1
click==7.0
xlrd==1.2.0
loguru==0.3.2
psutil==5.6.3
flask_cors==3.0.8
spacy==2.1.8
Flask-APScheduler==1.11.0
pyfiglet==0.8.post1
docx2python
msoffcrypto-tool==4.10.1
